+++ packages/core/src/date/performanceNow.ts
@@ -1,0 +1,14 @@
+import { useNow, NowOptions } from "./now";
+import { isBoolean } from "../utils";
+
+
+export function usePerformanceNow(options?: NowOptions) {
+  const refreshMs = options && options.refreshMs || 1000;
+  const sync = options && isBoolean(options.sync) ? options.sync : true;
+
+  return useNow({
+    refreshMs,
+    sync,
+    timeFn: () => performance.now()
+  })
+}