+++ packages/core/src/pagination/pagination.ts
@@ -1,5 +1,4 @@
-// import { Ref, computed, ref } from "@vue/reactivity";
-import { minMax, RefTyped, wrap } from "@vue-composable/core";
+import { isNumber, minMax, wrap, RefTyped } from "../utils";
 import { watch, Ref, computed, ref } from "@vue/runtime-core";
 
 type PaginationControl = () => void;
@@ -35,7 +34,7 @@ export function usePagination(options: PaginationOptions): PaginationResult {
       return _offset.value;
     },
     set(v) {
-      if (typeof v !== "number") {
+      if (!isNumber(v)) {
         /* istanbul ignore else */
         if (__DEV__) {
           console.warn(
@@ -53,7 +52,7 @@ export function usePagination(options: PaginationOptions): PaginationResult {
       return _currentPage.value;
     },
     set(v) {
-      if (typeof v !== "number") {
+      if (!isNumber(v)) {
         /* istanbul ignore else  */
         if (__DEV__) {
           console.warn(
@@ -73,7 +72,7 @@ export function usePagination(options: PaginationOptions): PaginationResult {
       return _pageSize.value;
     },
     set(v) {
-      if (typeof v !== "number") {
+      if (!isNumber(v)) {
         /* istanbul ignore else */
         if (__DEV__) {
           console.warn(