+++ packages/web/__tests__/utils.ts
@@ -1,14 +1,1 @@
 export { nextTick } from '@vue/runtime-core'
-// export const Vue: any = require('vue/dist/vue.common');
-
-// Vue.config.productionTip = false;
-// Vue.config.devtools = false;
-
-// export function nextTick<T>(callback: (this: T) => void, context?: T): void;
-// export function nextTick(): Promise<void>;
-// export function nextTick<T>(callback?: (this: T) => void, context?: T) {
-//     if(!callback) {
-//         return Vue.nextTick();
-//     }
-//     return Vue.nextTick(callback, context);
-// }