+++ packages/web/src/event/onScroll.ts
@@ -1,5 +1,5 @@
 import { Ref, ref } from "@vue/runtime-core";
-import { RefElement, wrap, RefTyped, useDebounce } from "@vue-composable/core";
+import { RefElement, wrap, RefTyped, useDebounce, isNumber } from "@vue-composable/core";
 import { useEvent, RemoveEventFunction } from "./event";
 
 export interface ScrollResult {
@@ -35,8 +35,7 @@ export function useOnScroll(
     scrollLeft.value = element.value.scrollLeft;
   };
 
-  const eventOptions = typeof options === "number" ? undefined : options;
-  const ms = typeof options === "number" ? options : wait;
+  const [eventOptions, ms] = isNumber(options) ? [undefined, options] : [options, wait];
 
   if (ms) {
     handler = useDebounce(handler, wait);