+++ rollup.config.js
@@ -139,7 +139,11 @@ function createConfig(output, plugins = [], config = {}) {
     external:
       isGlobalBuild || isRawESMBuild
         ? config.external || []
-        : knownExternals.concat(Object.keys(pkg.dependencies || [])),
+        : knownExternals.concat(
+            Object.keys(pkg.dependencies || []).concat(
+              Object.keys(pkg.peerDependencies || []) || []
+            )
+          ),
     plugins: [
       ...(config.plugins || []),
 
@@ -156,12 +160,7 @@ function createConfig(output, plugins = [], config = {}) {
       ),
       ...plugins
     ],
-    output,
-    onwarn: (msg, warn) => {
-      if (!/Circular/.test(msg)) {
-        warn(msg);
-      }
-    }
+    output
   };
 }
 
@@ -187,7 +186,9 @@ function createReplacePlugin(
     __RUNTIME_COMPILE__: isRuntimeCompileBuild,
     // support options?
     // the lean build drops options related code with buildOptions.lean: true
-    "process.env.NODE_ENV": !isProduction
+    "process.env.NODE_ENV": isBundlerESMBuild
+      ? `process.env.NODE_ENV`
+      : "'production'"
   });
 }